+++ lib/ui/profile/ProfileInfoEditPage.dart
@@ -51,7 +51,7 @@ class ProfileInfoEditPage extends StatefulWidget {
   State<StatefulWidget> createState() => ProfileInfoEditPageState();
 }
 
-class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<ProfileInfoEditPage> with NotificationsListener, WidgetsBindingObserver {
+class ProfileInfoEditPageState extends State<ProfileInfoEditPage> with NotificationsListener, WidgetsBindingObserver {
 
   late Auth2UserProfileFieldsVisibility _profileVisibility;
   late Uint8List? _pronunciationAudioData;
@@ -398,6 +398,9 @@ class ProfileInfoEditPageState extends ProfileDirectoryMyInfoBasePageState<Profi
       _fieldTextControllers[_ProfileField.middleName]?.text,
       _fieldTextControllers[_ProfileField.lastName]?.text,
     ]);
+
+    TextStyle? get nameTextStyle =>
+      Styles().textStyles.getTextStyleEx('widget.title.medium_large.fat', fontHeight: 0.85, textOverflow: TextOverflow.ellipsis);
 
   // Edit: Pronunciation